cloud_functions/etf_data_ingest/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
requests==2.*

